+++ src/main/java/com/xlongwei/light4j/util/PlateUtil.java
@@ -63,7 +63,7 @@ public class PlateUtil {
 		try{
 			ExcelUtil.readBySax(new BufferedInputStream(ConfigUtil.stream("chepaihao.xlsx")), 0, new RowHandler() {
 				@Override
-				public void handle(int i, int j, List<Object> list) {
+				public void handle(int i, long j, List<Object> list) {
 					if(j > 0) {
 						plates.add(list.toArray(new String[list.size()]));
 					}
